+++ scripts/lib/summons.js
@@ -132,12 +132,13 @@ export class Summons {
         await Summons.dismiss({trigger});
         return true;
     }
-    static async getSummonItem(name, updates, originItem, {flatAttack = false, flatDC = false, damageBonus = null, translate, identifier, damageFlat = null} = {}) {
+    static async getSummonItem(name, updates, originItem, {flatAttack = false, flatDC = false, damageBonus = null, translate, identifier, damageFlat = null, rules = 'legacy'} = {}) {
         let bonuses = (new Roll(originItem.actor.system.bonuses.rsak.attack + ' + 0', originItem.actor.getRollData()).evaluateSync({strict: false})).total;
         let prof = originItem.actor.system.attributes.prof;
         let abilityModifier = originItem.actor.system.abilities[originItem.abilityMod ?? originItem.actor.system.attributes?.spellcasting].mod;
         let attackBonus = bonuses + prof + abilityModifier;
-        let documentData = await compendiumUtils.getItemFromCompendium(constants.featurePacks.summonFeatures, name, {
+        let compendium = rules === 'modern' ? constants.modernPacks.summonFeatures : constants.featurePacks.summonFeatures;
+        let documentData = await compendiumUtils.getItemFromCompendium(compendium, name, {
             object: true, 
             getDescription: true, 
             translate,